backend/oneshotngo/requirements.txt
django==2.0.2
django-markdownx
django-extensions==1.9.9
djangorestframework==3.7.7
markdown==2.6.11
django-cors-headers
mysqlclient
